+++ test/fixtures/cookbooks/test/recipes/adoptopenjdk-8-openj9.rb
@@ -1,0 +1,6 @@
+node.default['java']['install_flavor'] = 'adoptopenjdk'
+node.default['java']['jdk_version'] = '8'
+
+include_recipe 'test::base'
+include_recipe 'java::default'
+include_recipe 'test::java_cert'